Renamings and more thinking about chip requirements
