+++ src/chip/mod.rs
@@ -1,0 +1,8 @@
+mod nand;
+pub use nand::NAND;
+
+mod chip;
+pub use chip::{Chip, Input, Output, Link };
+
+#[cfg(test)]
+mod tests;
