+++ src/chip/nand.rs
@@ -1,21 +1,27 @@
 
-pub struct NANDGate {
+pub struct NAND {
     pub input_a: u8,
     pub input_b: u8,
     pub vcc: u8,
+    pub gnd: u8,
 }
 
-impl NANDGate {
+impl NAND {
     pub fn new() -> Self {
         Self {
             input_a: 0,
             input_b: 0,
             vcc: 0,
+            gnd: 0,
         }
     }
 
     pub fn set_vcc(&mut self, vcc: u8) {
         self.vcc = vcc;
+    }
+
+    pub fn set_gnd(&mut self, gnd: u8) {
+        self.gnd = gnd;
     }
 
     pub fn set_inputs(&mut self, input_a: u8, input_b: u8) {
@@ -24,7 +30,7 @@ impl NANDGate {
     }
 
     pub fn output(&self) -> u8 {
-        if self.vcc == 0 {
+        if self.vcc == 0 || self.gnd == 1 {
             return 0;
         }
