+++ src/chip/tests/mod.rs
@@ -1,0 +1,5 @@
+#[cfg(test)]
+mod test_nand;
+
+#[cfg(test)]
+mod test_chip;
