--- src/component/mod.rs
@@ -1,8 +1,0 @@
-mod nand_gate;
-pub use nand_gate::NANDGate;
-
-mod component;
-pub use component::{Component, Input, Output, Link };
-
-#[cfg(test)]
-mod tests;
