--- src/component/tests/mod.rs
@@ -1,5 +1,0 @@
-#[cfg(test)]
-mod test_nand_gate;
-
-#[cfg(test)]
-mod test_component;
