+++ src/lib.rs
@@ -1,4 +1,4 @@
-pub mod component;
+pub mod chip;
 
 pub fn add(left: u64, right: u64) -> u64 {
     left + right
